requirements.txt
pyyaml
python-chess
